started assignment for Glue
