+++ Module 2 class 2/Assignment/mockdata.py
@@ -1,27 +1,29 @@
 import csv
 import random
-from datetime import datetime, timedelta
+from datetime import datetime
+
+# Function to read customer information from CSV file
+def read_customer_info(filename):
+    customer_info = {}
+    with open(filename, mode='r') as file:
+        reader = csv.DictReader(file)
+        for row in reader:
+            customer_info[row['name']] = {
+                'customer_id': int(row['customer_id']),
+                'debit_card_number': row['debit_card_number'],
+                'debit_card_type': row['debit_card_type'],
+                'bank_name': row['bank_name']
+            }
+    return customer_info
 
 # Function to generate mock data for daily transactions
-def generate_mock_transactions(num_transactions):
-    # Sample data for generating mock transactions
-    customer_names = ["Himanshu Bhatt", "Vaibhavi Nautiyal", "Naruto Uzumaki", "Sasuke Uchiha", "Tanjiro Kamado", "Itadori Sukuna"]
-    debit_card_types = ["Visa", "Mastercard", "American Express", "Discover"]
-    bank_names = ["Bank of America", "HDFC", "SBI", "Citibank", "Standard Chartered"]
-    
-    # Generate transactions
+def generate_mock_transactions(customer_info, num_transactions):
     transactions = []
     for _ in range(num_transactions):
-        customer_id = random.randint(1000, 9999)
-        name = random.choice(customer_names)
-        debit_card_number = ''.join(str(random.randint(0, 9)) for _ in range(16))
-        debit_card_type = random.choice(debit_card_types)
-        bank_name = random.choice(bank_names)
-        transaction_date = datetime.now().strftime('%Y-%m-%d')
-        amount_spend = round(random.uniform(10, 500), 2)
-        
-        transactions.append([customer_id, name, debit_card_number, debit_card_type, bank_name, transaction_date, amount_spend])
-    
+        for name, info in customer_info.items():
+            transaction_date = datetime(2024, 3, 23).strftime('%Y-%m-%d')
+            amount_spend = round(random.uniform(10, 500), 2)  # Varies for each transaction
+            transactions.append([info["customer_id"], name, info["debit_card_number"], info["debit_card_type"], info["bank_name"], transaction_date, amount_spend])
     return transactions
 
 # Function to save transactions data to a CSV file
@@ -31,13 +33,21 @@ def save_to_csv(data, filename):
         writer.writerow(['customer_id', 'name', 'debit_card_number', 'debit_card_type', 'bank_name', 'transaction_date', 'amount_spend'])
         writer.writerows(data)
 
-# Generate transactions for today
-num_transactions = 50  # You can change this number as per your requirement
-transactions = generate_mock_transactions(num_transactions)
+# Main function
+def main():
+    # Read customer information from CSV file
+    customer_info = read_customer_info('customer_records.csv')
+    
+    # Generate transactions for today
+    num_transactions = 20  # You can change this number as per your requirement
+    transactions = generate_mock_transactions(customer_info, num_transactions)
+    
+    # Save transactions to a CSV file       
+    today_date = datetime(2024, 3, 23).strftime('%Y-%m-%d') #datetime.now().strftime('%Y-%m-%d')
+    csv_filename = f"transactions_{today_date}.csv"
+    save_to_csv(transactions, csv_filename)
+    
+    print(f"Mock transactions for {today_date} have been generated and saved to '{csv_filename}'.")
 
-# Save transactions to a CSV file
-today_date = datetime.now().strftime('%Y-%m-%d')
-csv_filename = f"transactions_{today_date}.csv"
-save_to_csv(transactions, csv_filename)
-
-print(f"Mock transactions for {today_date} have been generated and saved to '{csv_filename}'.")
+if __name__ == "__main__":
+    main()
